Fix identifying wrong type in log messages

+++ utils-core/src/main/java/org/trasler/utils/lang/TypeUtils.java
@@ -71,7 +71,7 @@ public class TypeUtils {
                 }
                 catch (NumberFormatException e) {
                     // Fall through.
-                    logger.warn("Could not convert from:{} to:{}", value.getClass(), Double.class);
+                    logger.warn("Could not convert from:{} to:{}", value.getClass(), Integer.class);
                 }
             }
             case Boolean b -> {
@@ -106,7 +106,7 @@ public class TypeUtils {
                 }
                 catch (NumberFormatException e) {
                     // Fall through.
-                    logger.warn("Could not convert from:{} to:{}", value.getClass(), Double.class);
+                    logger.warn("Could not convert from:{} to:{}", value.getClass(), Long.class);
                 }
             }
             case Boolean b -> {
